Refactored theme message types (DH-17600)
